+++ pkg/dartdev/test/native_assets/build_test.dart
@@ -29,78 +29,59 @@ void main([List<String> args = const []]) async {
   final dartDevEntryScriptUri = resolveDartDevUri('bin/dartdev.dart');
 
   final bool fromDartdevSource = args.contains('--source');
-  final hostOS = Platform.operatingSystem;
-  final crossOS = Platform.isLinux ? 'macos' : 'linux';
-  for (final targetOS in [null, hostOS, crossOS]) {
-    String? osModifier;
-    if (targetOS == hostOS) {
-      osModifier = 'host';
-    } else if (targetOS == crossOS) {
-      osModifier = 'cross';
-    }
-    for (final verbose in [true, false]) {
-      final testModifier = [
-        '',
-        if (osModifier != null) osModifier,
-        if (verbose) 'verbose'
-      ].join(' ');
-      test('dart build$testModifier', timeout: longTimeout, () async {
-        await nativeAssetsTest('dart_app', (dartAppUri) async {
-          final bool expectCrossOSFailure = targetOS == crossOS;
-          final result = await runDart(
-            arguments: [
-              if (fromDartdevSource) dartDevEntryScriptUri.toFilePath(),
-              'build',
-              if (targetOS != null) ...[
-                '--target-os',
-                targetOS,
-              ],
-              if (verbose) '-v',
-              'bin/dart_app.dart',
-            ],
-            workingDirectory: dartAppUri,
-            logger: logger,
-            expectExitCodeZero: !expectCrossOSFailure,
-          );
-          if (expectCrossOSFailure) {
-            expect(result.stderr, contains(crossOSNotAllowedError('exe')));
-            expect(result.stderr, contains(hostOSMessage));
-            expect(result.stderr, contains(targetOSMessage(crossOS)));
-            expect(result.exitCode, 128);
-            return; // No executable to run.
-          }
-          if (verbose) {
-            expect(result.stdout, contains(usingTargetOSMessage));
-            expect(result.stdout, contains('build.dart'));
-          } else {
-            expect(result.stdout, isNot(contains('build.dart')));
-          }
 
-          final relativeExeUri = Uri.file('./bin/dart_app/dart_app.exe');
-          final absoluteExeUri = dartAppUri.resolveUri(relativeExeUri);
-          expect(await File.fromUri(absoluteExeUri).exists(), true);
-          await _withTempDir((tempUri) async {
-            // The link needs to have the same extension as the executable on
-            // Windows to be able to be executable.
-            final link = Link.fromUri(tempUri.resolve('my_link.exe'));
-            await link.create(absoluteExeUri.toFilePath());
-            for (final exeUri in [
-              absoluteExeUri,
-              relativeExeUri,
-              link.uri,
-            ]) {
-              final result = await runProcess(
-                executable: exeUri,
-                arguments: [],
-                workingDirectory: dartAppUri,
-                logger: logger,
-              );
-              expectDartAppStdout(result.stdout);
-            }
-          });
+  /// The relative uri from the package root to the app bundle.
+  final relativeBundleUri = Uri.directory(
+      './build/cli/${OS.current}-${Architecture.current}/bundle/');
+
+  for (final verbose in [true, false]) {
+    final testModifier = verbose ? ' verbose' : '';
+    test('dart build$testModifier', timeout: longTimeout, () async {
+      await nativeAssetsTest('dart_app', (dartAppUri) async {
+        final result = await runDart(
+          arguments: [
+            if (fromDartdevSource) dartDevEntryScriptUri.toFilePath(),
+            'build',
+            'cli',
+            if (verbose) '-v',
+          ],
+          workingDirectory: dartAppUri,
+          logger: logger,
+        );
+        if (verbose) {
+          expect(result.stdout, contains(usingTargetOSMessage));
+          expect(result.stdout, contains('build.dart'));
+        } else {
+          expect(result.stdout, isNot(contains('build.dart')));
+        }
+
+        final relativeExeUri = relativeBundleUri
+            .resolve('bin/')
+            .resolve(OS.current.executableFileName('dart_app'));
+        final absoluteExeUri = dartAppUri.resolveUri(relativeExeUri);
+        expect(await File.fromUri(absoluteExeUri).exists(), true);
+        await _withTempDir((tempUri) async {
+          // The link needs to have the same extension as the executable on
+          // Windows to be able to be executable.
+          final link = Link.fromUri(
+              tempUri.resolve(OS.current.executableFileName('my_link')));
+          await link.create(absoluteExeUri.toFilePath());
+          for (final exeUri in [
+            absoluteExeUri,
+            relativeExeUri,
+            link.uri,
+          ]) {
+            final result = await runProcess(
+              executable: exeUri,
+              arguments: [],
+              workingDirectory: dartAppUri,
+              logger: logger,
+            );
+            expectDartAppStdout(result.stdout);
+          }
         });
       });
-    }
+    });
   }
 
   test('dart build native assets build failure', timeout: longTimeout,
@@ -115,7 +96,7 @@ void main(List<String> args) {
       final result = await runDart(
         arguments: [
           'build',
-          'bin/dart_app.dart',
+          'cli',
         ],
         workingDirectory: dartAppUri,
         logger: logger,
@@ -136,7 +117,7 @@ void main(List<String> args) {
       final result = await runDart(
         arguments: [
           'build',
-          'bin/native_add_duplicate.dart',
+          'cli',
         ],
         workingDirectory: dartAppUri,
         logger: logger,
@@ -157,7 +138,7 @@ void main(List<String> args) {
       await runDart(
         arguments: [
           'build',
-          'bin/drop_dylib_link.dart',
+          'cli',
         ],
         workingDirectory: dartAppUri,
         logger: logger,
@@ -165,12 +146,12 @@ void main(List<String> args) {
       );
 
       // Check that the build directory exists
-      final directory =
-          Directory.fromUri(dartAppUri.resolve('bin/drop_dylib_link'));
-      expect(directory.existsSync(), true);
+      final libDirectory = Directory.fromUri(
+          dartAppUri.resolveUri(relativeBundleUri).resolve('lib/'));
+      expect(libDirectory.existsSync(), true);
 
       // Check that only one dylib is in the final application package
-      final buildFiles = directory.listSync(recursive: true);
+      final buildFiles = libDirectory.listSync(recursive: true);
       expect(
         buildFiles.where((file) => file.path.contains('add')),
         isNotEmpty,
@@ -187,7 +168,7 @@ void main(List<String> args) {
       final result = await runDart(
         arguments: [
           'build',
-          'bin/add_asset_link.dart',
+          'cli',
         ],
         workingDirectory: dartAppUri,
         logger: logger,
@@ -200,7 +181,7 @@ void main(List<String> args) {
 
       // Check that the build directory exists
       final directory =
-          Directory.fromUri(dartAppUri.resolve('bin/add_asset_link'));
+          Directory.fromUri(dartAppUri.resolveUri(relativeBundleUri));
       expect(directory.existsSync(), true);
       final dylib = OS.current.libraryFileName('add', DynamicLoadingBundled());
       expect(
@@ -216,8 +197,8 @@ void main(List<String> args) {
         arguments: [
           if (fromDartdevSource) dartDevEntryScriptUri.toFilePath(),
           'build',
-          'bin/dart_app.dart',
-          '.'
+          'cli',
+          '--output=.'
         ],
         workingDirectory: dartAppUri,
         logger: logger,
@@ -248,6 +229,8 @@ void main(List<String> args) {
           arguments: [
             '--enable-experiment=record-use',
             'build',
+            'cli',
+            '--target',
             'bin/$filename.dart',
           ],
           workingDirectory: dartAppUri,
@@ -255,26 +238,28 @@ void main(List<String> args) {
           expectExitCodeZero: true,
         );
 
+        final bundleDirectory =
+            Directory.fromUri(dartAppUri.resolveUri(relativeBundleUri));
         await runProcess(
-          executable: Uri.file('bin/$filename/$filename.exe'),
+          executable: bundleDirectory.uri
+              .resolve('bin/')
+              .resolve(OS.current.executableFileName(filename)),
           logger: logger,
           expectedExitCode: 0,
           throwOnUnexpectedExitCode: true,
           workingDirectory: dartAppUri,
         );
 
-        // The build directory exists
-        final shakeDirectory =
-            Directory.fromUri(dartAppUri.resolve('bin/$filename'));
-        expect(shakeDirectory.existsSync(), true);
+        // The build directory exists.
+        expect(bundleDirectory.existsSync(), true);
 
-        // The multiply asset has been treeshaken
+        // The multiply asset has been treeshaken.
         expect(
-          File.fromUri(shakeDirectory.uri.resolve('lib/$addLib')).existsSync(),
+          File.fromUri(bundleDirectory.uri.resolve('lib/$addLib')).existsSync(),
           true,
         );
         expect(
-          File.fromUri(shakeDirectory.uri.resolve('lib/$mulitplyLib'))
+          File.fromUri(bundleDirectory.uri.resolve('lib/$mulitplyLib'))
               .existsSync(),
           false,
         );
@@ -290,18 +275,18 @@ void main(List<String> args) {
         await runDart(
           arguments: [
             'build',
-            'bin/native_dynamic_linking.dart',
+            'cli',
           ],
           workingDirectory: packageUri,
           logger: logger,
         );
 
-        final outputDirectory =
-            Directory.fromUri(packageUri.resolve('bin/native_dynamic_linking'));
-        expect(outputDirectory.existsSync(), true);
+        final bundleDirectory =
+            Directory.fromUri(packageUri.resolveUri(relativeBundleUri));
+        expect(bundleDirectory.existsSync(), true);
 
         File dylibFile(String name) {
-          final libDirectoryUri = (outputDirectory.uri.resolve('lib/'));
+          final libDirectoryUri = (bundleDirectory.uri.resolve('lib/'));
           final dylibBasename =
               OS.current.libraryFileName(name, DynamicLoadingBundled());
           return File.fromUri(libDirectoryUri.resolve(dylibBasename));
@@ -312,7 +297,9 @@ void main(List<String> args) {
         expect(dylibFile('debug').existsSync(), true);
 
         final proccessResult = await runProcess(
-          executable: outputDirectory.uri.resolve('native_dynamic_linking.exe'),
+          executable: bundleDirectory.uri
+              .resolve('bin/')
+              .resolve(OS.current.executableFileName('native_dynamic_linking')),
           logger: logger,
           throwOnUnexpectedExitCode: true,
         );
@@ -331,18 +318,20 @@ void main(List<String> args) {
           await runDart(
             arguments: [
               'build',
-              'bin/user_defines.dart',
+              'cli',
             ],
             workingDirectory: packageUri,
             logger: logger,
           );
 
-          final outputDirectory =
-              Directory.fromUri(packageUri.resolve('bin/user_defines'));
-          expect(outputDirectory.existsSync(), true);
+          final bundleDirectory =
+              Directory.fromUri(packageUri.resolveUri(relativeBundleUri));
+          expect(bundleDirectory.existsSync(), true);
 
           final proccessResult = await runProcess(
-            executable: outputDirectory.uri.resolve('user_defines.exe'),
+            executable: bundleDirectory.uri
+                .resolve('bin/')
+                .resolve(OS.current.executableFileName('user_defines')),
             logger: logger,
             throwOnUnexpectedExitCode: true,
           );